re-organize the code
